+++ DSM-G02-PI3-2022-2-main/DSM-G02-PI3-2022-2-main/frontend/projeto-interdisciplinar/src/components/Header/header.css
@@ -53,5 +53,9 @@
 }
 
 .icone {
- 
+    position: absolute;
+    width: 50.2px;
+    height: 52px;
+    left: 1290px;
+    top: 12px;
 }
