+++ DSM-G02-PI3-2022-2-main/DSM-G02-PI3-2022-2-main/frontend/projeto-interdisciplinar/src/pages/Glossario/Glossario.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import Header from '../../components/Header/header';
+import './glossario.css';
+
+//Pagina Questoes da Aplicação
+export default function Glossario() {
+    return (
+        <div className='app'>
+            <Header />
+
+        </div>
+    )
+
+}
